// src/app/core/domain/enums/request-status.enum.ts
export enum RequestStatuses {
    Active = 'ACTIVE',
    Incoming = 'INCOMING',
    Pending = 'PENDING',
    Done = 'DONE',
}

export enum RequestStatuseDetails {
    Coming = 'COMING',
    Arrived = 'ARRIVED',
    Started = 'STARTED',
    Ended = 'ENDED',
    Cancelled = 'CANCELLED',
    Await = 'AWAIT'
}
  